fix(block_structure): reorganize block computation
Significant block information which was previously displayed as raw
integers is now displayed in hexadecimal. The header hash now contains
the entire block header, instead of just an ad-hoc computation of it.
The blocks are now asynchronously computed in sequence, rather than out
of order, which would theoretically allow new blocks to be hashed with
the prev_hash field intact.

Signed-off-by: Buckley Ross <brr322@lehigh.edu>
Change-Id: Ie8526e974e046d27af44acdad9e286a6b779f995